pyproject.toml
[tool.poetry]
name = "valeo-neuroerp"
version = "1.0.1"
description = "VALEO-NeuroERP System mit GENXAIS-Framework Integration"
authors = ["VALEO Team"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.115.14"
uvicorn = "0.24.0"
pymongo = "4.6.1"
pydantic = "2.11.7"
python-dotenv = "1.1.1"
requests = "2.32.4"
httpx = "0.28.1"
python-multipart = "0.0.9"
python-jose = "3.3.0"
passlib = "1.7.4"
bcrypt = "4.1.2"
aiofiles = "23.2.1"
jinja2 = "3.1.3"
starlette = "0.46.2"

[tool.poetry.group.dev.dependencies]
pytest = "8.1.1"
pytest-asyncio = "0.23.5"
black = "^24.1.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.13.2"

[tool.poetry.group.monitoring.dependencies]
prometheus-client = "^0.19.0"
structlog = "^24.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
asyncio_mode = "auto" 